@@ -1,12 +1,18 @@
 import React from 'react'
 import './App.css'
-import {AppBar, Button, Container, IconButton, LinearProgress, Toolbar, Typography} from '@material-ui/core'
-import {Menu} from '@material-ui/icons'
-import {TodolistsList} from '../features/TodolistsList/TodolistsList'
-import {ErrorSnackbar} from '../components/ErrorSnackbar/ErrorSnackbar'
-import {useSelector} from 'react-redux'
-import {AppRootStateType} from './store'
-import {RequestStatusType} from './app-reducer'
+import { TodolistsList } from '../features/TodolistsList/TodolistsList'
+import { useSelector } from 'react-redux'
+import { AppRootStateType } from './store'
+import { RequestStatusType } from './app-reducer'
+import AppBar from '@mui/material/AppBar';
+import Toolbar from '@mui/material/Toolbar';
+import IconButton from '@mui/material/IconButton';
+import Typography from '@mui/material/Typography';
+import Button from '@mui/material/Button';
+import Container from '@mui/material/Container';
+import LinearProgress from '@mui/material/LinearProgress';
+import { Menu } from '@mui/icons-material';
+import { ErrorSnackbar } from '../components/ErrorSnackbar/ErrorSnackbar'
 
 type PropsType = {
     demo?: boolean
@@ -16,7 +22,7 @@ function App({demo = false}: PropsType) {
     const status = useSelector<AppRootStateType, RequestStatusType>((state) => state.app.status)
     return (
         <div className="App">
-            <ErrorSnackbar />
+            <ErrorSnackbar/>
             <AppBar position="static">
                 <Toolbar>
                     <IconButton edge="start" color="inherit" aria-label="menu">
@@ -27,7 +33,7 @@ function App({demo = false}: PropsType) {
                     </Typography>
                     <Button color="inherit">Login</Button>
                 </Toolbar>
-             { status === 'loading' &&  <LinearProgress /> }
+                {status === 'loading' && <LinearProgress/>}
             </AppBar>
             <Container fixed>
                 <TodolistsList demo={demo}/>
