@@ -1,8 +1,8 @@
-import {tasksReducer} from '../features/TodolistsList/tasks-reducer';
-import {todolistsReducer} from '../features/TodolistsList/todolists-reducer';
-import {applyMiddleware, combineReducers, createStore} from 'redux'
+import { tasksReducer } from '../features/TodolistsList/tasks-reducer';
+import { todolistsReducer } from '../features/TodolistsList/todolists-reducer';
+import { applyMiddleware, combineReducers, createStore } from 'redux'
 import thunkMiddleware from 'redux-thunk'
-import {appReducer} from './app-reducer'
+import { appReducer } from './app-reducer'
 
 // объединяя reducer-ы с помощью combineReducers,
 // мы задаём структуру нашего единственного объекта-состояния
